Set up class/module empty lines Rubocop rules
Empty lines inside class definitions. No empty lines inside module
definitions. To support this, extracted out proxy `Periods` and
`Timeline` objects to handle directional concerns.
