+++ lib/biz/core_ext/date.rb
@@ -1,11 +1,9 @@
 module Biz
   module CoreExt
     module Date
-
       def business_day?
         Biz.dates.active?(self)
       end
-
     end
   end
 end
