+++ lib/biz/core_ext/fixnum.rb
@@ -1,11 +1,9 @@
 module Biz
   module CoreExt
     module Fixnum
-
       Calculation::ForDuration::UNITS.each do |unit|
         define_method("business_#{unit}") { Biz.time(self, unit) }
       end
-
     end
   end
 end
