+++ lib/biz/core_ext/time.rb
@@ -1,11 +1,9 @@
 module Biz
   module CoreExt
     module Time
-
       def business_hours?
         Biz.in_hours?(self)
       end
-
     end
   end
 end
