+++ lib/biz/day_time.rb
@@ -2,10 +2,8 @@ module Biz
   class DayTime
 
     module Timestamp
-
       FORMAT  = '%02d:%02d'
       PATTERN = /\A(?<hour>\d{2}):(?<minute>\d{2})(:?(?<second>\d{2}))?\Z/
-
     end
 
     include Comparable
