+++ lib/biz/periods.rb
@@ -1,22 +1,5 @@
 module Biz
-  class Periods
-
-    def initialize(schedule)
-      @schedule = schedule
-    end
-
-    def after(origin)
-      After.new(schedule, origin)
-    end
-
-    def before(origin)
-      Before.new(schedule, origin)
-    end
-
-    protected
-
-    attr_reader :schedule
-
+  module Periods
   end
 end
 
@@ -24,3 +7,4 @@ require 'biz/periods/abstract'
 
 require 'biz/periods/after'
 require 'biz/periods/before'
+require 'biz/periods/proxy'
