+++ lib/biz/timeline.rb
@@ -1,22 +1,5 @@
 module Biz
-  class Timeline
-
-    def initialize(periods)
-      @periods = periods
-    end
-
-    def forward
-      Forward.new(periods)
-    end
-
-    def backward
-      Backward.new(periods)
-    end
-
-    protected
-
-    attr_reader :periods
-
+  module Timeline
   end
 end
 
@@ -24,3 +7,4 @@ require 'biz/timeline/abstract'
 
 require 'biz/timeline/forward'
 require 'biz/timeline/backward'
+require 'biz/timeline/proxy'
