+++ lib/biz/timeline/abstract.rb
@@ -1,5 +1,5 @@
 module Biz
-  class Timeline
+  module Timeline
     class Abstract
 
       def initialize(periods)
