+++ lib/biz/version.rb
@@ -1,5 +1,3 @@
 module Biz
-
   VERSION = '1.2.1'
-
 end
